experiment with imgui and classes refactoring
