+++ Src/Projects/shader_GPU_Particles/ReadMe.md
@@ -59,16 +59,16 @@ Particle System
 
 
  13.12.2013
-    Первое применение частицам, это симуляция эффектов для поведения автомобиля. Дым из трубы, пыль из-под колес.
-  Интересно было бы сделать также пост эффект рефракции, как на автомобильной дороге в жаркую погоду.
-    На данный момент прилагающимся силам не хватает турбулентности, шума. Надо будет разобраться как сделать подобный
-эффект, возможно с помощью 3д текстуры шума. Причем текстура может быть векторной, чтобы оттуда извлекать сразу значения
-для velocity частиц в поле.
-	Другой момент связан с симуляцией частиц. На данный момент проигрываются они произвольно, а алгоритм рассчета обычный Эйлера.
-Можно попробовать добавить алгоритм Рунга-Кутта 3 или 4й степени или же ввести на данный момент минимальный шаг рассчета, чтобы
-делать все вычисления с одинаковым шагом и тогда результат будет предсказуемым и его можно будет проигрывать во времени.
-	Также возможность кеширования весьма интересна тем, чтобы получить нужный вид симуляции и обмениваться с Майкой или наоборот,
-учитывая что формат кеша мне уже известен (спасибо автору из Диснея, кто поделился исходниками чтения формата)
+    ГЏГҐГ°ГўГ®ГҐ ГЇГ°ГЁГ¬ГҐГ­ГҐГ­ГЁГҐ Г·Г Г±ГІГЁГ¶Г Г¬, ГЅГІГ® Г±ГЁГ¬ГіГ«ГїГ¶ГЁГї ГЅГґГґГҐГЄГІГ®Гў Г¤Г«Гї ГЇГ®ГўГҐГ¤ГҐГ­ГЁГї Г ГўГІГ®Г¬Г®ГЎГЁГ«Гї. Г„Г»Г¬ ГЁГ§ ГІГ°ГіГЎГ», ГЇГ»Г«Гј ГЁГ§-ГЇГ®Г¤ ГЄГ®Г«ГҐГ±.
+  Г€Г­ГІГҐГ°ГҐГ±Г­Г® ГЎГ»Г«Г® ГЎГ» Г±Г¤ГҐГ«Г ГІГј ГІГ ГЄГ¦ГҐ ГЇГ®Г±ГІ ГЅГґГґГҐГЄГІ Г°ГҐГґГ°Г ГЄГ¶ГЁГЁ, ГЄГ ГЄ Г­Г  Г ГўГІГ®Г¬Г®ГЎГЁГ«ГјГ­Г®Г© Г¤Г®Г°Г®ГЈГҐ Гў Г¦Г Г°ГЄГіГѕ ГЇГ®ГЈГ®Г¤Гі.
+    ГЌГ  Г¤Г Г­Г­Г»Г© Г¬Г®Г¬ГҐГ­ГІ ГЇГ°ГЁГ«Г ГЈГ ГѕГ№ГЁГ¬Г±Гї Г±ГЁГ«Г Г¬ Г­ГҐ ГµГўГ ГІГ ГҐГІ ГІГіГ°ГЎГіГ«ГҐГ­ГІГ­Г®Г±ГІГЁ, ГёГіГ¬Г . ГЌГ Г¤Г® ГЎГіГ¤ГҐГІ Г°Г Г§Г®ГЎГ°Г ГІГјГ±Гї ГЄГ ГЄ Г±Г¤ГҐГ«Г ГІГј ГЇГ®Г¤Г®ГЎГ­Г»Г©
+ГЅГґГґГҐГЄГІ, ГўГ®Г§Г¬Г®Г¦Г­Г® Г± ГЇГ®Г¬Г®Г№ГјГѕ 3Г¤ ГІГҐГЄГ±ГІГіГ°Г» ГёГіГ¬Г . ГЏГ°ГЁГ·ГҐГ¬ ГІГҐГЄГ±ГІГіГ°Г  Г¬Г®Г¦ГҐГІ ГЎГ»ГІГј ГўГҐГЄГІГ®Г°Г­Г®Г©, Г·ГІГ®ГЎГ» Г®ГІГІГіГ¤Г  ГЁГ§ГўГ«ГҐГЄГ ГІГј Г±Г°Г Г§Гі Г§Г­Г Г·ГҐГ­ГЁГї
+Г¤Г«Гї velocity Г·Г Г±ГІГЁГ¶ Гў ГЇГ®Г«ГҐ.
+	Г„Г°ГіГЈГ®Г© Г¬Г®Г¬ГҐГ­ГІ Г±ГўГїГ§Г Г­ Г± Г±ГЁГ¬ГіГ«ГїГ¶ГЁГҐГ© Г·Г Г±ГІГЁГ¶. ГЌГ  Г¤Г Г­Г­Г»Г© Г¬Г®Г¬ГҐГ­ГІ ГЇГ°Г®ГЁГЈГ°Г»ГўГ ГѕГІГ±Гї Г®Г­ГЁ ГЇГ°Г®ГЁГ§ГўГ®Г«ГјГ­Г®, Г  Г Г«ГЈГ®Г°ГЁГІГ¬ Г°Г Г±Г±Г·ГҐГІГ  Г®ГЎГ»Г·Г­Г»Г© ГќГ©Г«ГҐГ°Г .
+ГЊГ®Г¦Г­Г® ГЇГ®ГЇГ°Г®ГЎГ®ГўГ ГІГј Г¤Г®ГЎГ ГўГЁГІГј Г Г«ГЈГ®Г°ГЁГІГ¬ ГђГіГ­ГЈГ -ГЉГіГІГІГ  3 ГЁГ«ГЁ 4Г© Г±ГІГҐГЇГҐГ­ГЁ ГЁГ«ГЁ Г¦ГҐ ГўГўГҐГ±ГІГЁ Г­Г  Г¤Г Г­Г­Г»Г© Г¬Г®Г¬ГҐГ­ГІ Г¬ГЁГ­ГЁГ¬Г Г«ГјГ­Г»Г© ГёГ ГЈ Г°Г Г±Г±Г·ГҐГІГ , Г·ГІГ®ГЎГ»
+Г¤ГҐГ«Г ГІГј ГўГ±ГҐ ГўГ»Г·ГЁГ±Г«ГҐГ­ГЁГї Г± Г®Г¤ГЁГ­Г ГЄГ®ГўГ»Г¬ ГёГ ГЈГ®Г¬ ГЁ ГІГ®ГЈГ¤Г  Г°ГҐГ§ГіГ«ГјГІГ ГІ ГЎГіГ¤ГҐГІ ГЇГ°ГҐГ¤Г±ГЄГ Г§ГіГҐГ¬Г»Г¬ ГЁ ГҐГЈГ® Г¬Г®Г¦Г­Г® ГЎГіГ¤ГҐГІ ГЇГ°Г®ГЁГЈГ°Г»ГўГ ГІГј ГўГ® ГўГ°ГҐГ¬ГҐГ­ГЁ.
+	Г’Г ГЄГ¦ГҐ ГўГ®Г§Г¬Г®Г¦Г­Г®Г±ГІГј ГЄГҐГёГЁГ°Г®ГўГ Г­ГЁГї ГўГҐГ±ГјГ¬Г  ГЁГ­ГІГҐГ°ГҐГ±Г­Г  ГІГҐГ¬, Г·ГІГ®ГЎГ» ГЇГ®Г«ГіГ·ГЁГІГј Г­ГіГ¦Г­Г»Г© ГўГЁГ¤ Г±ГЁГ¬ГіГ«ГїГ¶ГЁГЁ ГЁ Г®ГЎГ¬ГҐГ­ГЁГўГ ГІГјГ±Гї Г± ГЊГ Г©ГЄГ®Г© ГЁГ«ГЁ Г­Г Г®ГЎГ®Г°Г®ГІ,
+ГіГ·ГЁГІГ»ГўГ Гї Г·ГІГ® ГґГ®Г°Г¬Г ГІ ГЄГҐГёГ  Г¬Г­ГҐ ГіГ¦ГҐ ГЁГ§ГўГҐГ±ГІГҐГ­ (Г±ГЇГ Г±ГЁГЎГ® Г ГўГІГ®Г°Гі ГЁГ§ Г„ГЁГ±Г­ГҐГї, ГЄГІГ® ГЇГ®Г¤ГҐГ«ГЁГ«Г±Гї ГЁГ±ГµГ®Г¤Г­ГЁГЄГ Г¬ГЁ Г·ГІГҐГ­ГЁГї ГґГ®Г°Г¬Г ГІГ )
  
   + add support of transparency modes (old task - sort particles and draw after everything)
